[feature] Add ChartViewGroup to ImageCanvas
Added ChartViewGroup to ImageCanvas to display chart data outside of the image
