+++ src/components/DataGetHandler.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from "react"
-import { dataListElement, multiDataData } from "../context/ToolsBarContext"
+import { dataListElement } from "../context/ToolsBarContext"
 import { useImages } from "../hooks/useImages"
 import PixelLocation from "./PixelLocation"
 import useToolBar from "../hooks/useToolBar"
@@ -10,7 +10,6 @@ import { useLoader } from "../hooks/useLoader"
 import { listen } from "@tauri-apps/api/event"
 import { event } from "@tauri-apps/api"
 import SelectedArea from "./SelectedArea"
-import ChartView from "./ChartView"
 import AreaLocation from "./AreaLocation"
 
 interface selectedArea {
@@ -169,14 +168,6 @@ export default function DataGetHandler() {
     })
     setMouseSelectedArea({ x1: null, y1: null, x2: null, y2: null })
   }
-
-  const coordsToText = (coords: { x: number, y: number } | { x1: number, y1: number, x2: number, y2: number }) => {
-    if ('x' in coords) {
-      return `${coords.x}, ${coords.y}`
-    }
-    return `${Math.min(coords.x1, coords.x2)}, ${Math.min(coords.y1, coords.y2)} - ${Math.max(coords.x1, coords.x2)}, ${Math.max(coords.y1, coords.y2)}`
-  }
-
   return (
 
     currentImage && <div ref={selfRef} className="h-full w-full absolute">
@@ -229,32 +220,7 @@ export default function DataGetHandler() {
             })
           )
         }
-        </div>
-        {
-          currentDataList.length > 0 && (
-            currentDataList.map((data: dataListElement) => (
-              <ChartView
-                key={`chart-${data.id}`}
-                id={data.id}
-                parentRef={selfRef}
-                title={ `${data.type[0].toUpperCase() + data.type.slice(1)} | ${coordsToText(data.coords)}`}
-                dataList={data.type === 'area' ? [{
-                  title: "min",
-                  data: (data.data as multiDataData).min,
-                  style: { color: '#339AF0', lineStyle: 1 }
-                }, {
-                  title: "max",
-                  data: (data.data as multiDataData).max,
-                  style: { color: '#FA5252', lineStyle: 1 }
-                }, {
-                  title: "mean",
-                  data: (data.data as multiDataData).mean,
-                }] : 
-                [{ data: data.data as number[] }]}
-              />
-            ))
-          )
-        }
+      </div>
     </div>
   )
 }
